backporting: leave backport/author PRs alone
[ upstream commit 396135588213aaabf9de60d6bed5d041b590d3eb ]

This label means that the backport will be performed by the author, and
thus the backporting automation should ignore it.

It's okay to not filter by backport-done/<branch> because these labels
should not be present at the same time. (Notably, if resiliency was the
intention, we should already be filtering backport-pending too.)

Signed-off-by: David Bimmler <david.bimmler@isovalent.com>
Signed-off-by: Quentin Monnet <quentin@isovalent.com>